+++ frontend/src/modules/forum/configs/config.js
@@ -1,0 +1,1 @@
+export const ROOT_URL = 'http://localhost:8000'
