+++ frontend/src/modules/forum/views/colleges/index.js
