+++ frontend/src/modules/forum/views/course/index.js
