+++ frontend/src/modules/forum/views/courses/actions.js
@@ -1,3 +1,5 @@
+import {ROOT_URL} from "../../configs/config"
+
 export const GET_COURSES = 'get_courses';
 export const GOT_COURSES = 'got_courses';
 
@@ -14,6 +16,6 @@ export const getCourses = (collegeId) => {
 };
 
 function fetchCourses(collegeId) {
-    return fetch(`http://localhost:8000/api/forum/courses?collegeid=${collegeId}`)
+    return fetch(`${ROOT_URL}/api/forum/courses?collegeid=${collegeId}`)
         .then((response) => (response.json()))
 }
