+++ frontend/src/modules/forum/views/courses/index.js
