+++ frontend/src/modules/forum/views/courses/reducers.js
