+++ frontend/src/modules/forum/views/main/actions.js
@@ -1,3 +1,5 @@
+import {ROOT_URL} from "../../configs/config"
+
 export const GET_SUBSCRIPTIONS = 'get_subscriptions';
 export const GOT_SUBSCRIPTIONS = 'got_subscriptions';
 
@@ -14,6 +16,6 @@ export const getSubscriptions = (userId, token) => {
 };
 
 function fetchSubscriptions(userId, token) {
-    return fetch(`http://localhost:8000/api/forum/subscriptions?uid=${userId}&token=${token}`)
+    return fetch(`${ROOT_URL}/api/forum/subscriptions?uid=${userId}&token=${token}`)
         .then((response) => (response.json()))
 }
