+++ frontend/src/modules/forum/views/main/index.js
